IGNITE-20449 Introduce ThreadUtils class
